requirements.txt
fastapi
uvicorn
opencv-python
numpy
python-multipart
python-jose[cryptography]
